Move a group dynamically

move_group2 moves a group to a different position from within a shiny server function. The parameter from refers to a panel-id of a panel within the group you want to move. Likewise to refers to a panel-id of a panel within the group you want to select as to. The difference between move_group2() and move_group() is that move_group2() selects both from and to by panel-id, whereas move_group() selects by group-id.

Usage

move_group2(
  dock_id,
  from,
  to,
  position = NULL,
  session = shiny::getDefaultReactiveDomain()
)

Arguments

dock_id

Dock unique id. When using modules the namespace is automatically added.

from

Panel-id of a panel within the group that should be moved.

to

Panel-id of a panel within the group you want as a to.

position

Group position options: one of "left", "right", "top", "bottom", "center".

session

shiny session object. See https://dockview.dev/docs/api/dockview/panelApi/.

Value

This function is called for its side effect. It sends a message to JavaScript through the current websocket connection, leveraging the shiny session object.