get dock

get dock

get dock panels

get dock panels ids

get dock active group

get dock grid

get dock groups

get dock groups ids

get dock groups panels

save a dock

restore a dock

Usage

get_dock(dock_id, session = getDefaultReactiveDomain())

get_panels(dock_id, session = getDefaultReactiveDomain())

get_panels_ids(dock_id, session = getDefaultReactiveDomain())

get_active_group(dock_id, session = getDefaultReactiveDomain())

get_grid(dock_id, session = getDefaultReactiveDomain())

get_groups(dock_id, session = getDefaultReactiveDomain())

get_groups_ids(dock_id, session = getDefaultReactiveDomain())

get_groups_panels(dock_id, session = getDefaultReactiveDomain())

save_dock(dock_id, session = getDefaultReactiveDomain())

restore_dock(dock_id, data, session = getDefaultReactiveDomain())

Arguments

dock_id

Dock unique id. When using modules the namespace is automatically added.

session

shiny session object.

data

Data representing a serialised dock object.

Value

get_dock returns a list of 3 elements:

grid: a list representing the dock layout.

panels: a list having the same structure as panel() composing the dock.

activeGroup: the current active group (a string).

Each other function allows to deep dive into the returned value of get_dock(). get_panels() returns the panels element of get_dock(). get_panels_ids() returns a character vector containing all panel ids from get_panels(). get_active_group() extracts the activeGroup component of get_dock() as a string. get_grid() returns the grid element of get_dock() which is a list. get_groups() returns a list of panel groups from get_grid(). get_groups_ids() returns a character vector of groups ids from get_groups(). get_groups_panels() returns a list of character vector containing the ids of each panel within each group. save_dock() and restore_dock() are used for their side effect to allow to respectively serialise and restore a dock object.

Note

Only works with server side functions like add_panel. Don't call it from the UI.