Create a dock view widget

Creates an interactive dock view widget that enables flexible layout management with draggable, resizable, and dockable panels. This is a wrapper around the dockview.dev JavaScript library, providing a powerful interface for creating IDE-like layouts in Shiny applications or R Markdown documents.

Usage

dock_view(
  panels,
  ...,
  theme = c("light-spaced", "light", "abyss", "abyss-spaced", "dark", "vs", "dracula",
    "replit"),
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

panels

Widget configuration. Slot for panel.

...

Other options. See https://dockview.dev/docs/api/dockview/options/.

theme

Theme. One of c("abyss", "dark", "light", "vs", "dracula", "replit").

width

Widget width.

height

Widget height.

elementId

When used outside Shiny.

Value

An HTML widget object.

Examples in Shinylive

example-1

Open in Shinylive